+++ package.json
@@ -39,8 +39,8 @@
     "element-ready": "^2.1.0",
     "escape-goat": "^1.1.0",
     "github-injection": "^0.3.0",
-    "select-dom": "^4.1.1",
     "hosted-git-info": "^2.5.0",
+    "select-dom": "^4.1.1",
     "webext-dynamic-content-scripts": "^2.0.1"
   }
 }
